Add other role and fix minor bugs

+++ app/Http/Controllers/ClassificationController.php
@@ -2,13 +2,14 @@
 
 namespace App\Http\Controllers;
 
-use App\Http\Requests\StoreClassificationRequest;
-use App\Http\Requests\UpdateClassificationRequest;
-use App\Models\Classification;
-use Illuminate\Contracts\View\View;
-use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use App\Models\Classification;
+use Illuminate\Contracts\View\View;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Http\RedirectResponse;
+use App\Http\Requests\StoreClassificationRequest;
+use App\Http\Requests\UpdateClassificationRequest;
 
 class ClassificationController extends Controller
 {
@@ -20,10 +21,15 @@ class ClassificationController extends Controller
      */
     public function index(Request $request): View
     {
-        return view('pages.reference.classification', [
-            'data' => Classification::render($request->search),
-            'search' => $request->search,
-        ]);
+        if (Auth::user()->role === 'Admin' || Auth::user()->role === 'Ketua P3MP') {
+            return view('pages.reference.classification', [
+                'data' => Classification::render($request->search),
+                'search' => $request->search,
+            ]);
+        } else {
+            abort(403);
+        }
+
     }
 
     /**
@@ -34,11 +40,15 @@ class ClassificationController extends Controller
      */
     public function store(StoreClassificationRequest $request): RedirectResponse
     {
-        try {
-            Classification::create($request->validated());
-            return back()->with('success', __('menu.general.success'));
-        } catch (\Throwable $exception) {
-            return back()->with('error', $exception->getMessage());
+        if (Auth::user()->role === 'Admin' || Auth::user()->role === 'Ketua P3MP') {
+            try {
+                Classification::create($request->validated());
+                return back()->with('success', __('menu.general.success'));
+            } catch (\Throwable $exception) {
+                return back()->with('error', $exception->getMessage());
+            }
+        } else {
+            abort(403);
         }
     }
 
@@ -51,11 +61,15 @@ class ClassificationController extends Controller
      */
     public function update(UpdateClassificationRequest $request, Classification $classification): RedirectResponse
     {
-        try {
-            $classification->update($request->validated());
-            return back()->with('success', __('menu.general.success'));
-        } catch (\Throwable $exception) {
-            return back()->with('error', $exception->getMessage());
+        if (Auth::user()->role === 'Admin' || Auth::user()->role === 'Ketua P3MP') {
+            try {
+                $classification->update($request->validated());
+                return back()->with('success', __('menu.general.success'));
+            } catch (\Throwable $exception) {
+                return back()->with('error', $exception->getMessage());
+            }
+        } else {
+            abort(403);
         }
     }
 
@@ -67,11 +81,15 @@ class ClassificationController extends Controller
      */
     public function destroy(Classification $classification): RedirectResponse
     {
-        try {
-            $classification->delete();
-            return back()->with('success', __('menu.general.success'));
-        } catch (\Throwable $exception) {
-            return back()->with('error', $exception->getMessage());
+        if (Auth::user()->role === 'Admin' || Auth::user()->role === 'Ketua P3MP') {
+            try {
+                $classification->delete();
+                return back()->with('success', __('menu.general.success'));
+            } catch (\Throwable $exception) {
+                return back()->with('error', $exception->getMessage());
+            }
+        } else {
+            abort(403);
         }
     }
 }
